Create an Outputs class to handle writing outputs to file
