+++ src/model/model.hpp
@@ -22,12 +22,12 @@
 #include "elements/springs/springs_input.hpp"
 #include "mesh_connectivity.hpp"
 #include "node.hpp"
+#include "outputs.hpp"
 #include "solver/solver.hpp"
 #include "state/state.hpp"
 #include "step/step_parameters.hpp"
 #include "system/beams/update_node_state.hpp"
 #include "types.hpp"
-#include "utilities/netcdf/node_state_writer.hpp"
 
 namespace openturbine {
 
@@ -467,82 +467,19 @@ public:
     // Outputs
     //--------------------------------------------------------------------------
 
-    /// @brief Set up the NetCDF output writer
-    void SetOutputWriter(const std::string& output_file, size_t num_nodes) {
-        this->output_writer_ = std::make_unique<util::NodeStateWriter>(output_file, true, num_nodes);
+    /// @brief Set up outputs configuration
+    /// @param output_file Path to output file
+    /// @param num_nodes Number of nodes in the model
+    void SetupOutputs(const std::string& output_file, size_t num_nodes) {
+        this->outputs_ = std::make_unique<Outputs>(output_file, num_nodes);
     }
 
-    void WriteNodeOutputsAtTimestep(State& state, size_t timestep) const {
-        if (!this->output_writer_) {
-            return;
+    /// @brief Write node state outputs at the specified timestep
+    /// @param timestep Current timestep index
+    void WriteOutputsAtTimestep(State& state, size_t timestep) const {
+        if (this->outputs_) {
+            outputs_->WriteNodeOutputsAtTimestep(state, timestep);
         }
-
-        const size_t num_nodes = state.num_system_nodes;
-        std::vector<double> x(num_nodes);
-        std::vector<double> y(num_nodes);
-        std::vector<double> z(num_nodes);
-        std::vector<double> i(num_nodes);
-        std::vector<double> j(num_nodes);
-        std::vector<double> k(num_nodes);
-        std::vector<double> w(num_nodes);
-
-        // Create host mirrors of device views
-        auto host_q = Kokkos::create_mirror_view(state.q);
-        auto host_v = Kokkos::create_mirror_view(state.v);
-        auto host_vd = Kokkos::create_mirror_view(state.vd);
-        auto host_x = Kokkos::create_mirror_view(state.x);
-
-        // Copy data from device to host
-        Kokkos::deep_copy(host_q, state.q);
-        Kokkos::deep_copy(host_v, state.v);
-        Kokkos::deep_copy(host_vd, state.vd);
-        Kokkos::deep_copy(host_x, state.x);
-
-        // Position data
-        for (size_t node = 0; node < num_nodes; ++node) {
-            x[node] = host_x(node, 0);
-            y[node] = host_x(node, 1);
-            z[node] = host_x(node, 2);
-            w[node] = host_x(node, 3);
-            i[node] = host_x(node, 4);
-            j[node] = host_x(node, 5);
-            k[node] = host_x(node, 6);
-        }
-        this->output_writer_->WriteStateDataAtTimestep(timestep, "x", x, y, z, i, j, k, w);
-
-        // Displacement data
-        for (size_t node = 0; node < num_nodes; ++node) {
-            x[node] = host_q(node, 0);
-            y[node] = host_q(node, 1);
-            z[node] = host_q(node, 2);
-            w[node] = host_q(node, 3);
-            i[node] = host_q(node, 4);
-            j[node] = host_q(node, 5);
-            k[node] = host_q(node, 6);
-        }
-        this->output_writer_->WriteStateDataAtTimestep(timestep, "u", x, y, z, i, j, k, w);
-
-        // Velocity data
-        for (size_t node = 0; node < num_nodes; ++node) {
-            x[node] = host_v(node, 0);
-            y[node] = host_v(node, 1);
-            z[node] = host_v(node, 2);
-            i[node] = host_v(node, 3);
-            j[node] = host_v(node, 4);
-            k[node] = host_v(node, 5);
-        }
-        this->output_writer_->WriteStateDataAtTimestep(timestep, "v", x, y, z, i, j, k);
-
-        // Acceleration data
-        for (size_t node = 0; node < num_nodes; ++node) {
-            x[node] = host_vd(node, 0);
-            y[node] = host_vd(node, 1);
-            z[node] = host_vd(node, 2);
-            i[node] = host_vd(node, 3);
-            j[node] = host_vd(node, 4);
-            k[node] = host_vd(node, 5);
-        }
-        this->output_writer_->WriteStateDataAtTimestep(timestep, "a", x, y, z, i, j, k);
     }
 
 private:
@@ -553,7 +490,7 @@ private:
     std::vector<SpringElement> spring_elements_;  //< Spring elements in the model
     std::vector<Constraint> constraints_;         //< Constraints in the model
     MeshConnectivity mesh_connectivity_;  //< Mesh connectivity tracking element-node relationships
-    std::unique_ptr<util::NodeStateWriter> output_writer_;  //< NetCDF output writer
+    std::unique_ptr<Outputs> outputs_;    //< Outputs
 };
 
 }  // namespace openturbine
